refactor common parts into common package
