+++ app/src/main/java/com/shearer/jetmoviedb/features/movie/MovieModule.kt
@@ -1,12 +1,12 @@
 package com.shearer.jetmoviedb.features.movie
 
 import androidx.room.Room
-import com.shearer.jetmoviedb.features.movie.db.MovieDb
-import com.shearer.jetmoviedb.features.movie.interactor.MovieInteractor
-import com.shearer.jetmoviedb.features.movie.interactor.MovieInteractorDefault
+import com.shearer.jetmoviedb.features.movie.common.db.MovieDb
+import com.shearer.jetmoviedb.features.movie.common.interactor.MovieInteractor
+import com.shearer.jetmoviedb.features.movie.common.interactor.MovieInteractorDefault
 import com.shearer.jetmoviedb.features.movie.list.MovieListViewModel
-import com.shearer.jetmoviedb.features.movie.repository.MovieRepository
-import com.shearer.jetmoviedb.features.movie.repository.MovieRepositoryDefault
+import com.shearer.jetmoviedb.features.movie.common.repository.MovieRepository
+import com.shearer.jetmoviedb.features.movie.common.repository.MovieRepositoryDefault
 import com.shearer.jetmoviedb.shared.http.RestServiceDefault
 import org.koin.androidx.viewmodel.ext.koin.viewModel
 import org.koin.dsl.module.module
