+++ app/src/main/java/com/shearer/jetmoviedb/features/movie/common/db/MovieDb.kt
@@ -1,10 +1,10 @@
-package com.shearer.jetmoviedb.features.movie.db
+package com.shearer.jetmoviedb.features.movie.common.db
 
 import android.content.Context
 import androidx.room.Database
 import androidx.room.Room
 import androidx.room.RoomDatabase
-import com.shearer.jetmoviedb.features.movie.domain.Movie
+import com.shearer.jetmoviedb.features.movie.common.domain.Movie
 
 @Database(entities = [Movie::class], version = 1, exportSchema = false)
 abstract class MovieDb : RoomDatabase() {
