+++ app/src/main/java/com/shearer/jetmoviedb/features/movie/common/domain/MovieResults.kt
@@ -1,3 +1,3 @@
-package com.shearer.jetmoviedb.features.movie.domain
+package com.shearer.jetmoviedb.features.movie.common.domain
 
 data class MovieResults(val page: Int, val totalResults: Int, val totalPages: Int, val movies: List<Movie>)
