+++ app/src/main/java/com/shearer/jetmoviedb/features/movie/common/repository/MovieDbApi.kt
@@ -1,7 +1,7 @@
-package com.shearer.jetmoviedb.features.movie.repository
+package com.shearer.jetmoviedb.features.movie.common.repository
 
-import com.shearer.jetmoviedb.features.movie.repository.dto.GenreResultDto
-import com.shearer.jetmoviedb.features.movie.repository.dto.MovieResultDto
+import com.shearer.jetmoviedb.features.movie.common.repository.dto.GenreResultDto
+import com.shearer.jetmoviedb.features.movie.common.repository.dto.MovieResultDto
 import io.reactivex.Single
 import retrofit2.http.GET
 import retrofit2.http.Query
