+++ app/src/main/java/com/shearer/jetmoviedb/features/movie/common/repository/dto/GenreDto.kt
@@ -1,0 +1,3 @@
+package com.shearer.jetmoviedb.features.movie.common.repository.dto
+
+data class GenreDto(val id: Int, val name: String)
