+++ app/src/main/java/com/shearer/jetmoviedb/features/movie/common/repository/dto/GenreResultDto.kt
@@ -1,4 +1,4 @@
-package com.shearer.jetmoviedb.features.movie.repository.dto
+package com.shearer.jetmoviedb.features.movie.common.repository.dto
 
 data class GenreResultDto(val genres: List<GenreDto>) {
