--- app/src/main/java/com/shearer/jetmoviedb/features/movie/repository/dto/GenreDto.kt
@@ -1,3 +1,0 @@
-package com.shearer.jetmoviedb.features.movie.repository.dto
-
-data class GenreDto(val id: Int, val name: String)
