+++ app/src/test/java/com/shearer/jetmoviedb/features/movie/common/interactor/MovieInteractorDefaultTest.kt
@@ -1,14 +1,14 @@
-package com.shearer.jetmoviedb.features.movie.interactor
+package com.shearer.jetmoviedb.features.movie.common.interactor
 
 import androidx.paging.DataSource
 import com.nhaarman.mockito_kotlin.any
 import com.nhaarman.mockito_kotlin.doReturn
 import com.nhaarman.mockito_kotlin.mock
 import com.shearer.jetmoviedb.createPopularMovies
-import com.shearer.jetmoviedb.features.movie.db.MovieDao
-import com.shearer.jetmoviedb.features.movie.db.MovieDb
-import com.shearer.jetmoviedb.features.movie.domain.Movie
-import com.shearer.jetmoviedb.features.movie.repository.MovieRepository
+import com.shearer.jetmoviedb.features.movie.common.db.MovieDao
+import com.shearer.jetmoviedb.features.movie.common.db.MovieDb
+import com.shearer.jetmoviedb.features.movie.common.domain.Movie
+import com.shearer.jetmoviedb.features.movie.common.repository.MovieRepository
 import io.reactivex.Single
 import io.reactivex.disposables.CompositeDisposable
 
@@ -18,7 +18,7 @@ class MovieInteractorDefaultTest {
 
     private val source = mock<DataSource.Factory<Int, Movie>>()
     private val movieDao = mock<MovieDao> {
-        on { movies() } doReturn source
+        on { moviesByType(any()) } doReturn source
     }
     private val dbDao = mock<MovieDb> {
         on { movies() } doReturn movieDao
