+++ app/src/test/java/com/shearer/jetmoviedb/features/movie/common/repository/dto/GenreResultDtoTest.kt
@@ -1,4 +1,4 @@
-package com.shearer.jetmoviedb.features.movie.repository.dto
+package com.shearer.jetmoviedb.features.movie.common.repository.dto
 
 import com.google.common.truth.Truth.assertThat
 import com.shearer.jetmoviedb.createGenreDto
