+++ app/src/test/java/com/shearer/jetmoviedb/features/movie/common/repository/dto/MovieResultDtoTest.kt
@@ -1,4 +1,4 @@
-package com.shearer.jetmoviedb.features.movie.repository.dto
+package com.shearer.jetmoviedb.features.movie.common.repository.dto
 
 import com.google.common.truth.Truth.assertThat
 import com.shearer.jetmoviedb.createGenreDto
@@ -7,12 +7,13 @@ import org.junit.Test
 
 class MovieResultDtoTest {
 
+    private val type = "type"
     private val genres = createGenreDto().toGenres()
     private val movieResultDto = createPopularMoviesDto()
 
     @Test
     fun toMovies_convertsDtoToDomain() {
-        movieResultDto.toMovies(genres).run {
+        movieResultDto.toMovies(genres, type).run {
             assertThat(page).isEqualTo(1)
             assertThat(totalResults).isEqualTo(19831)
             assertThat(totalPages).isEqualTo(992)
